@@ -48,8 +48,8 @@
   },
   "dependencies": {
     "@vercel/commerce": "workspace:*",
-    "lodash.debounce": "^4.0.8",
-    "cookie": "^0.4.1"
+    "cookie": "^0.4.1",
+    "lodash.debounce": "^4.0.8"
   },
   "peerDependencies": {
     "next": "^13",
@@ -60,13 +60,13 @@
     "@taskr/clear": "^1.1.0",
     "@taskr/esnext": "^1.1.0",
     "@taskr/watch": "^1.1.0",
+    "@types/cookie": "^0.4.1",
     "@types/lodash.debounce": "^4.0.6",
     "@types/node": "^17.0.8",
-    "@types/react": "^18.0.14",
-    "@types/cookie": "^0.4.1",
     "@types/node-fetch": "^2.6.2",
+    "@types/react": "^18.0.14",
     "lint-staged": "^12.1.7",
-    "next": "^13.0.6",
+    "next": "^13.2.4",
     "prettier": "^2.5.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
