@@ -75,7 +75,7 @@
     "dotenv": "^16.0.1",
     "graphql": "^16.0.0",
     "lint-staged": "^13.0.3",
-    "next": "^12.2.1",
+    "next": "^12.3.4",
     "prettier": "^2.7.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
