@@ -68,7 +68,7 @@
     "@types/node-fetch": "^2.6.2",
     "@types/react": "^18.0.14",
     "lint-staged": "^12.1.7",
-    "next": "^13.0.6",
+    "next": "^13.2.4",
     "prettier": "^2.5.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
