@@ -1,6 +1,8 @@
+'use client'
+
 import cn from 'clsx'
 import { useTheme } from 'next-themes'
-import { useRouter } from 'next/router'
+import { usePathname, useRouter } from 'next/navigation'
 import { Moon, Sun } from '@components/icons'
 import s from './CustomerMenuContent.module.css'
 import useLogout from '@framework/auth/use-logout'
@@ -27,7 +29,7 @@ const LINKS = [
 export default function CustomerMenuContent() {
   const router = useRouter()
   const logout = useLogout()
-  const { pathname } = useRouter()
+  const pathname = usePathname()
   const { theme, setTheme } = useTheme()
 
   function handleClick(_: React.MouseEvent<HTMLAnchorElement>, href: string) {
