@@ -1,4 +1,6 @@
-import { FC, useRef, useEffect } from 'react'
+'use client'
+
+import { FC } from 'react'
 import { useUserAvatar } from '@lib/hooks/useUserAvatar'
 
 interface Props {
@@ -7,12 +9,10 @@ interface Props {
 }
 
 const Avatar: FC<Props> = ({}) => {
-  let ref = useRef() as React.MutableRefObject<HTMLInputElement>
   let { userAvatar } = useUserAvatar()
 
   return (
     <div
-      ref={ref}
       style={{ backgroundImage: userAvatar }}
       className="inline-block h-8 w-8 rounded-full border-2 border-primary hover:border-secondary focus:border-secondary transition-colors ease-linear"
     >
